deleting movie bucket list
