+++ lib/movie_bucket_list/send_to_firebase.dart
@@ -1,0 +1,8 @@
+
+import 'package:flutter/material.dart';
+
+// Import the firebase_core and cloud_firestore plugin
+import 'package:firebase_core/firebase_core.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+
+class SendToFirebase {}
